requirements.txt
appnope==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
bleach==3.3.0
cairocffi==1.2.0
cffi==1.14.5
cycler==0.10.0
Cython==0.29.22
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.3
importlib-metadata==3.7.3
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.3.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.4
mistune==0.8.4
natsort==7.0.1
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
ncls==0.0.57
nest-asyncio==1.5.1
networkx==2.5
notebook==6.2.0
numpy==1.20.1
ordered-set==4.0.2
packaging==20.9
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.8.1
pyparsing==2.4.7
pyranges==0.0.93
pyrle==0.0.32
pyrsistent==0.17.3
python-dateutil==2.8.1
python-igraph==0.8.0
pytz==2021.1
pyzmq==22.0.3
qtconsole==5.0.3
QtPy==1.9.0
scikit-learn==0.24.1
scipy==1.6.2
seaborn==0.11.1
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
sorted-nearest==0.0.32
tabulate==0.8.9
terminado==0.9.4
testpath==0.4.4
texttable==1.6.3
threadpoolctl==2.1.0
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.4.1
